// src/app/pages/home-page/interfaces/Questionnaire.ts
import { RegistredUser } from "./RegistredUser";

export interface Questionnaire{
    id: number;
    question1: boolean;
    question2: boolean;
    question3: boolean;
    question4: boolean;
    question5: boolean;
    question6: boolean;
    question7: boolean;
    question8: boolean;
    question9: boolean;
    question10: boolean;
    question11: boolean;
    question12: boolean;
    question13: boolean;
    question14: boolean;
    question15: boolean;
    question16: boolean;
    question17: boolean;
    question18: boolean;
    question19: boolean;
    question20: boolean;
    question21: boolean;
    question22: boolean;
    question23: boolean;
    question24: boolean;
    question25: boolean;
    question26: boolean;
    date: Date;
    user: RegistredUser;
}